agent/requirements.txt
google-generativeai
fastapi
uvicorn
python-dotenv